refactor: 예외 코드 수정 및 Gemini 예외 처리 로직 변경
