+++ backend/src/main/java/com/f12/moitz/infrastructure/client/gemini/GoogleGeminiClient.java
@@ -7,6 +7,7 @@ import com.f12.moitz.common.error.exception.RetryableApiException;
 import com.f12.moitz.infrastructure.PromptGenerator;
 import com.f12.moitz.application.dto.RecommendedLocationResponse;
 import com.f12.moitz.infrastructure.client.gemini.dto.RecommendedPlaceResponses;
+import com.f12.moitz.infrastructure.client.gemini.utils.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.genai.Client;
@@ -34,6 +35,7 @@ public class GoogleGeminiClient {
 
     private static final String GEMINI_MODEL = "gemini-2.0-flash";
 
+    private final JsonParser jsonParser;
     private final Client geminiClient;
     private final ObjectMapper objectMapper;
 
@@ -91,14 +93,17 @@ public class GoogleGeminiClient {
             if (e.code() == HttpStatus.TOO_MANY_REQUESTS.value() || message.contains("exceeded your current quota")) {
                 throw new ExternalApiException(ExternalApiErrorCode.EXCEEDED_GEMINI_API_TOKEN_QUOTA, e.getMessage());
             }
-            if (e.code() == HttpStatus.SERVICE_UNAVAILABLE.value()) {
-                throw new ExternalApiException(ExternalApiErrorCode.GEMINI_API_SERVER_UNAVAILABLE, e.getMessage());
+            if (e.code() == HttpStatus.SERVICE_UNAVAILABLE.value() || message.contains("try again later")) {
+                throw new RetryableApiException(ExternalApiErrorCode.GEMINI_API_SERVER_UNAVAILABLE, e.getMessage());
+            }
+            if (e.code() == HttpStatus.INTERNAL_SERVER_ERROR.value()) {
+                throw new RetryableApiException(ExternalApiErrorCode.GEMINI_API_SERVER_UNAVAILABLE, e.getMessage());
             }
             throw new ExternalApiException(ExternalApiErrorCode.INVALID_GEMINI_API_RESPONSE, e.getMessage());
         }
     }
 
-    protected GenerateContentResponse generate(final List<Content> contents, final GenerateContentConfig config) {
+    public GenerateContentResponse generate(final List<Content> contents, final GenerateContentConfig config) {
         final GenerateContentResponse generateContentResponse = geminiClient.models.generateContent(
                 GEMINI_MODEL,
                 contents,
@@ -124,8 +129,8 @@ public class GoogleGeminiClient {
                 throw new RetryableApiException(ExternalApiErrorCode.INVALID_GEMINI_RESPONSE_FORMAT);
             }
 
-            String cleanedText = cleanJsonResponse(originalText);
-            if (!isValidJson(cleanedText)) {
+            String cleanedText = jsonParser.cleanJsonResponse(originalText);
+            if (!jsonParser.isValidJson(cleanedText)) {
                 log.error("유효하지 않은 JSON 형식: {}", cleanedText);
                 throw new RetryableApiException(ExternalApiErrorCode.INVALID_GEMINI_RESPONSE_FORMAT);
             }
@@ -135,119 +140,6 @@ public class GoogleGeminiClient {
             log.error("Gemini 응답 파싱 중 오류 발생", e);
             throw new RetryableApiException(ExternalApiErrorCode.INVALID_GEMINI_RESPONSE_FORMAT);
         }
-    }
-
-    /**
-     * JSON 응답에서 마크다운 코드 블록과 불필요한 문자 제거
-     */
-    private String cleanJsonResponse(String text) {
-        if (text == null) {
-            return "";
-        }
-
-        // 1. 마크다운 코드 블록 제거 (```json, ```, ``` 등)
-        String cleaned = text.replaceAll("```json\\s*", "")
-                .replaceAll("```\\s*$", "")
-                .replaceAll("^```\\s*", "")
-                .trim();
-
-        // 2. 앞뒤 공백 및 개행 문자 정리
-        cleaned = cleaned.trim();
-
-        // 3. JSON이 아닌 텍스트가 앞에 있는 경우 처리
-        int jsonStart = findJsonStart(cleaned);
-        if (jsonStart > 0) {
-            cleaned = cleaned.substring(jsonStart);
-        }
-
-        // 4. JSON이 아닌 텍스트가 뒤에 있는 경우 처리
-        int jsonEnd = findJsonEnd(cleaned);
-        if (jsonEnd > 0 && jsonEnd < cleaned.length()) {
-            cleaned = cleaned.substring(0, jsonEnd + 1);
-        }
-
-        return cleaned.trim();
-    }
-
-    /**
-     * JSON 시작 위치 찾기 ({ 또는 [)
-     */
-    private int findJsonStart(String text) {
-        for (int i = 0; i < text.length(); i++) {
-            char c = text.charAt(i);
-            if (c == '{' || c == '[') {
-                return i;
-            }
-        }
-        return 0;
-    }
-
-    /**
-     * JSON 끝 위치 찾기 (} 또는 ])
-     */
-    private int findJsonEnd(String text) {
-        int braceCount = 0;
-        int bracketCount = 0;
-        boolean inString = false;
-        boolean escaped = false;
-
-        for (int i = 0; i < text.length(); i++) {
-            char c = text.charAt(i);
-
-            if (escaped) {
-                escaped = false;
-                continue;
-            }
-
-            if (c == '\\') {
-                escaped = true;
-                continue;
-            }
-
-            if (c == '"' && !escaped) {
-                inString = !inString;
-                continue;
-            }
-
-            if (inString) {
-                continue;
-            }
-
-            switch (c) {
-                case '{':
-                    braceCount++;
-                    break;
-                case '}':
-                    braceCount--;
-                    if (braceCount == 0 && bracketCount == 0) {
-                        return i;
-                    }
-                    break;
-                case '[':
-                    bracketCount++;
-                    break;
-                case ']':
-                    bracketCount--;
-                    if (braceCount == 0 && bracketCount == 0) {
-                        return i;
-                    }
-                    break;
-            }
-        }
-        return text.length() - 1;
-    }
-
-    /**
-     * 기본적인 JSON 유효성 검사
-     */
-    private boolean isValidJson(String text) {
-        if (text == null || text.trim().isEmpty()) {
-            return false;
-        }
-
-        String trimmed = text.trim();
-        return (trimmed.startsWith("{") && trimmed.endsWith("}")) ||
-                (trimmed.startsWith("[") && trimmed.endsWith("]"));
     }
 
     private <T> T readValue(final String content, final Class<T> valueType) {
